finally figured out why the test was acting strange on page load there are two fetches that happen now one that fetches the default preset and one that fetches all the user's stored presets in the db. what happened was some perfect way how the jest test was written caused a fetch mock impl leak into the next test where I needed a different response mocked but was getting unintended response in the succeeding test Which was causing an undesired render on the app inside the test where a container should have only had 1 child element but continued to have 3 children, 2 of which were buttons that were only supposed to render in the previous test.
I at first thought it was a state problem, and then mounted tree
cleanup problem, but then actually turned out to be a leaky mock impl
of fetch from a previous test being used causing the extra child elements
to render. Solved by moving the test to it's own file and have it's own scope
where it can use it's own impl of fetch and not get a leaky mock that
doesn't clear with any of the reset/clear/whatever methods that jest
provides. It seems these functions don't really work the way they seem to work or the way the documentation says it works, there is an issue that has been open since 2018 on this exact problem, and is still open to this day 2/4/2022. Who knows when it will be solved. For now we can at least be aware of it and avoid it as best as we can. DJ-Viking ~cheers